Rename typings to types
