+++ codex-ui/src/vue/button/index.ts
@@ -1,4 +1,4 @@
 import Button from './Button.vue'
-export * from './Button.typings.js';
+export * from './Button.types.js';
 
 export { Button };
