+++ codex-ui/src/vue/form/index.ts
@@ -1,4 +1,4 @@
-export * from './Field.typings.js'
+export * from './Field.types.js'
 import Field from './Field.vue';
 
 export { Field };
