+++ codex-ui/src/vue/input/index.ts
@@ -1,4 +1,4 @@
-export * from './Input.typings.js';
+export * from './Input.types.js';
 import Input from './Input.vue';
 
 export { Input };
